Configured Selection & Examination UI
Configured 2D Canvas elements to be The Selection and Examination UI. Once the vehicle is selected, the user will be able to scale, rotate, and position the prefab freely. There will be a UI Button for the user to interact with to enter Examine Mode.

Once in Examine Mode, the user can place the vehicle in Drive/Hover mode, power on/off the vehicle, rotate the vehicle to achieve their desired viewing angle, and can exit, bringing them back to the vehicle placement selection menu.

BUG TO FIX--> Once the vehicle is enters and exits examine mode, something breaks. When the vehicle is selected again, the Examine Button does not appear and if you try to reposition the vehicle, the scale resets.