Examinable Mechanic Test 1 & 2
Test one focused on putting the CyberStrider Prefab into Examination Mode. Which was done successfully.

Test two focused on exiting the CyberStrider prefab once in examination mode. Test was unsuccessful as the prefab was not placed back properly. Seems to be an issue with how the cache was set up with the original parent reference.  Doing more research and will test again.