Examinable.Test3
Test 3 on the examinable mechanic has the Examine Button appear when the prefab is selected.

Upon the selection of the examination button, the user is then entered into Examine Mode.

When the user enters examination mode the examination button is disabled and the exit button is enabled.

The exit button leads the user back to the prefabs onSelection() state.

Read more on the custom Examine Mechanic Here: https://medium.com/@jordantkay/list/custom-mechanic-ar-examine-2fddaa893ce8