PowerMode.Test1
Test1 of the Power Mode function introduces the ability to turn the vehicle prefab on/off.

Test1 focuses on turning the vehicle's material emissions on/off (setting emission color from white <on> to black <off>)

Created PowerSettings script to hold all references to visual effects that are needing to be turned on/off.

Test1 creates Target Material reference, and logic to switch between on & off.